requirements.txt
asgiref==3.2.10
catboost==0.24.3
Django==3.0
djangorestframework==3.10.0
joblib==1.0.0
lightgbm==2.3.0
pandas==1.2.1
pip==20.0.2
scikit-learn==0.24.1
scipy==1.4.1
xgboost==1.2.0
setuptools==49.6.0

torch==1.6.0